NextGen RPC-O
This change converts the RPC-O repository to simply consume OSA and get
out of the way.

Many of our extra capabilities have been moved to external repos and are
consumable as external dependencies. This change minimizes how we produce
OpenStack clouds and allows us to approach some of our value-added
solutions in different ways.

To start:
* Anything not core to the product has been removed.
* The structure of the repo has been modified allowing us to better understand
  what we have
* Tools we no longer maintain and are not required for our OpenStack
  deployment solution have been removed.
* Variables specific to auxiliary tooling have been truncated.
* The tight coupling between OpenStack-Ansible and RPC-O has been
  removed.
* The markdown documents within the repo have been updated for accuracy.

Future:
* Testing that is no longer core to the project will need to be removed.
* Development focus will need to change from internal to upstream
  or to our value-added solutions.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>